ref(stats-detectors): Move classes for better imports
