+++ src/sentry/statistical_detectors/base.py
@@ -1,0 +1,49 @@
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from datetime import datetime
+from enum import Enum
+from typing import Any, Mapping
+
+
+class TrendType(Enum):
+    Regressed = "regressed"
+    Improved = "improved"
+    Unchanged = "unchanged"
+
+
+@dataclass(frozen=True)
+class DetectorPayload:
+    project_id: int
+    group: str | int
+    fingerprint: str
+    count: float
+    value: float
+    timestamp: datetime
+
+
+@dataclass(frozen=True)
+class DetectorState(ABC):
+    @classmethod
+    @abstractmethod
+    def from_redis_dict(cls, data: Any) -> DetectorState:
+        ...
+
+    @abstractmethod
+    def to_redis_dict(self) -> Mapping[str | bytes, bytes | float | int | str]:
+        ...
+
+    @abstractmethod
+    def should_auto_resolve(self, target: float, rel_threshold: float) -> bool:
+        ...
+
+    @classmethod
+    @abstractmethod
+    def empty(cls) -> DetectorState:
+        ...
+
+
+@dataclass(frozen=True)
+class DetectorConfig(ABC):
+    ...
