+++ src/sentry/statistical_detectors/redis.py
@@ -1,30 +1,26 @@
 from __future__ import annotations
 
-from enum import Enum
 from typing import List, Mapping
 
 from django.conf import settings
 from sentry_redis_tools.clients import RedisCluster, StrictRedis
 
-from sentry.statistical_detectors.detector import DetectorPayload, DetectorStore
+from sentry.models.statistical_detectors import RegressionType
+from sentry.statistical_detectors.base import DetectorPayload
+from sentry.statistical_detectors.store import DetectorStore
 from sentry.utils import redis
 
 STATE_TTL = 24 * 60 * 60  # 1 day TTL
 
 
-class DetectorType(Enum):
-    ENDPOINT = "e"
-    FUNCTION = "f"
-
-
 class RedisDetectorStore(DetectorStore):
     def __init__(
         self,
-        detector_type: DetectorType,
+        regression_type: RegressionType,
         client: RedisCluster | StrictRedis | None = None,
         ttl=STATE_TTL,
     ):
-        self.detector_type = detector_type
+        self.regression_type = regression_type
         self.ttl = ttl
         self._client: RedisCluster | StrictRedis | None = None
 
@@ -65,7 +61,9 @@ class RedisDetectorStore(DetectorStore):
             pipeline.execute()
 
     def make_key(self, payload: DetectorPayload):
-        return f"sd:p:{payload.project_id}:{self.detector_type.value}:{payload.fingerprint}"
+        return (
+            f"sd:p:{payload.project_id}:{self.regression_type.abbreviate()}:{payload.fingerprint}"
+        )
 
     @staticmethod
     def get_redis_client() -> RedisCluster | StrictRedis:
